// src/api/generated/schemas/updateLocationRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * cid-api
 * LDAP 로그인 + JWT 인가 API
 * OpenAPI spec version: v1
 */

export interface UpdateLocationRequest {
  /**
   * 사이트 명
   * @minLength 0
   * @maxLength 50
   */
  locSiteNm?: string;
  /**
   * 층 명
   * @minLength 0
   * @maxLength 20
   */
  locFloorNm?: string;
  /**
   * LOCATION 타입 (IDC/OFFICE 등)
   * @minLength 0
   * @maxLength 20
   */
  locTpCd?: string;
  /**
   * LOCATION 설명
   * @minLength 0
   * @maxLength 200
   */
  locDescp?: string;
}